[Gold V] Title: 퇴사 2, Time: 672 ms, Memory: 316052 KB -BaekjoonHub

+++ 백준/Gold/15486. 퇴사 2/퇴사 2.java
@@ -9,7 +9,7 @@ public class Main {
         int n = Integer.parseInt(br.readLine());
         int[] T = new int[n+1];
         int[] P = new int[n+1];
-        long[] D = new long[n+2];
+        int[] D = new int[n+2];
 
         for (int i = 1; i <= n; i++) {
             st = new StringTokenizer(br.readLine());
